dashboard.py
import streamlit as st
import pandas as pd
# import principles_calculator # Assuma que você tem uma função de leitura de dados

st.set_page_config(layout="wide")
st.title("🛡️ Dashboard de Harmonia AGI-GAIA-TECHNE")

st.sidebar.header("Filtros de Tempo")
# ... (Adicionar filtros de data)

st.header("Índice de Harmonia (Logos-Ethos)")
# fig_harmony = principles_calculator.get_harmony_plot()
# st.plotly_chart(fig_harmony)

st.subheader("Últimas Violações Éticas (Firewall)")
# df_violations = principles_calculator.get_violations()
# st.dataframe(df_violations)
